+++ packages/create-app/template/+next+fuma-docs-mdx+static/lib/layout.shared.tsx
@@ -12,5 +12,6 @@ export function baseOptions(): BaseLayoutProps {
     nav: {
       title: 'My App',
     },
+    githubUrl: `https://github.com/${gitConfig.user}/${gitConfig.repo}`,
   };
 }