+++ packages/create-app/template/waku/src/lib/source.ts
@@ -7,3 +7,12 @@ export const source = loader({
   baseUrl: '/docs',
   plugins: [lucideIconsPlugin()],
 });
+
+export function getPageImage(slugs: string[]) {
+  const segments = [...slugs, 'image.webp'];
+
+  return {
+    segments,
+    url: `/og/docs/${segments.join('/')}`,
+  };
+}