+++ packages/create-app/template/waku/src/pages/docs/[...slugs].tsx
@@ -2,6 +2,7 @@ import { source } from '@/lib/source';
 import { PageProps } from 'waku/router';
 import defaultMdxComponents from 'fumadocs-ui/mdx';
 import { DocsBody, DocsDescription, DocsPage, DocsTitle } from 'fumadocs-ui/layouts/docs/page';
+import { getPageImage } from '@/lib/source';
 
 export default function DocPage({ slugs }: PageProps<'/docs/[...slugs]'>) {
   const page = source.getPage(slugs);
@@ -20,6 +21,7 @@ export default function DocPage({ slugs }: PageProps<'/docs/[...slugs]'>) {
   const MDX = page.data.body;
   return (
     <DocsPage toc={page.data.toc}>
+      <meta property="og:image" content={getPageImage(slugs).url} />
       <DocsTitle>{page.data.title}</DocsTitle>
       <DocsDescription>{page.data.description}</DocsDescription>
       <DocsBody>