+++ packages/create-app/template/waku/waku.config.ts
@@ -11,6 +11,9 @@ export default defineConfig({
     optimizeDeps: {
       exclude: ['fumadocs-ui', 'fumadocs-core'],
     },
+    ssr: {
+      external: ['@takumi-rs/image-response'],
+    },
 
     plugins: [tailwindcss(), mdx(MdxConfig), tsconfigPaths()],
   } satisfies UserConfig as Config['vite'],