+++ packages/mdx/src/loaders/mdx/remark-include.ts
@@ -4,7 +4,7 @@ import type { Code, Node, Root, RootContent } from 'mdast';
 import * as path from 'node:path';
 import * as fs from 'node:fs/promises';
 import { fumaMatter } from '@/utils/fuma-matter';
-import type { MdxJsxFlowElement, MdxJsxTextElement } from 'mdast-util-mdx-jsx';
+import type { MdxJsxFlowElement, MdxJsxTextElement } from 'mdast-util-mdx';
 import { remarkHeading } from 'fumadocs-core/mdx-plugins';
 import { VFile } from 'vfile';
 import type { Directives } from 'mdast-util-directive';