+++ packages/obsidian/src/remark/remark-block-id.ts
@@ -2,7 +2,7 @@ import { visit } from 'unist-util-visit';
 import type { Transformer } from 'unified';
 import type { Root } from 'mdast';
 import { replace } from '@/utils/mdast-replace';
-import type { MdxJsxFlowElement } from 'mdast-util-mdx-jsx';
+import type { MdxJsxFlowElement } from 'mdast-util-mdx';
 
 const Regex = /(?<!\\)\^(?<block_id>\w+)$/m;