+++ packages/doc-gen/CHANGELOG.md
@@ -1,5 +1,14 @@
 # fumadocs-docgen
 
+## 3.0.7
+
+### Patch Changes
+
+- 1a614de: enforce MDX stringifier by default
+- Updated dependencies [1a614de]
+- Updated dependencies [6ab6692]
+  - fumadocs-core@16.6.5
+
 ## 3.0.6
 
 ### Patch Changes